Promote MountedDrives to top-level internal type
MountedDrives is extracted from its nested position inside LockDirectory
into its own file with no logic changes.
